requirements.txt
crewai
crewai-tools
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-community
PyPDF2
pymongo
click
colorama
python-dotenv
requests
semgrep
pydantic
rich
pyyaml
openai
anthropic
google-generativeai